Extract git ignore implementation
