+++ cli/dstack/core/repo/local.py
@@ -1,8 +1,6 @@
-import fnmatch
 import tarfile
-from itertools import zip_longest
 from pathlib import Path
-from typing import Any, BinaryIO, Dict, List, Optional
+from typing import Any, BinaryIO, Dict, Optional
 
 from typing_extensions import Literal
 
@@ -10,6 +8,7 @@ from dstack.core.repo.base import Repo, RepoData, RepoInfo, RepoRef
 from dstack.utils.common import PathLike
 from dstack.utils.escape import escape_head
 from dstack.utils.hash import get_sha256, slugify
+from dstack.utils.ignore import GitIgnore
 from dstack.utils.workflows import load_workflows
 
 
@@ -58,79 +57,8 @@ class LocalRepo(Repo):
         return load_workflows(Path(self.repo_data.repo_dir) / ".dstack")
 
 
-class TarIgnore:
-    def __init__(
-        self, root_dir: PathLike, ignore_files: List[str] = None, globs: List[str] = None
-    ):
-        self.root_dir = Path(root_dir)
-        self.ignore_files = (
-            ignore_files
-            if ignore_files is not None
-            else [".gitignore", ".git/info/exclude", ".dstackignore"]
-        )
-        self.ignore_globs: Dict[str, List[str]] = {"": globs or []}
-
-    def load_ignore_file(self, path: str, ignore_file: Path):
-        if path not in self.ignore_globs:
-            self.ignore_globs[path] = []
-        with ignore_file.open("r") as f:
-            for line in f:
-                line = self.rstrip(line.rstrip("\n")).rstrip("/")
-                line = line.replace("\\ ", " ")
-                if line.startswith("#") or not line:
-                    continue
-                self.ignore_globs[path].append(line)
-
-    @staticmethod
-    def rstrip(value: str) -> str:
-        end = len(value) - 1
-        while end >= 0:
-            if not value[end].isspace():
-                break
-            if end > 0 and value[end - 1] == "\\":
-                break  # escaped space
-            end -= 1
-        else:
-            return ""
-        return value[: end + 1]
-
-    @staticmethod
-    def fnmatch(name: str, pattern: str, sep="/") -> bool:
-        if pattern.startswith(sep):
-            name = sep + name
-        for n, p in zip_longest(
-            reversed(name.split(sep)), reversed(pattern.split(sep)), fillvalue=None
-        ):
-            if p == "**":
-                raise NotImplementedError()
-            if p is None:
-                return True
-            if n is None or not fnmatch.fnmatch(n, p):
-                return False
-        return True
-
-    def ignore(self, path: str, sep="/") -> bool:
-        if not path:
-            return False
-        tokens = (sep + path).split(sep)
-        for i in range(1, len(tokens)):
-            parent = sep.join(tokens[:-i])
-            globs = self.ignore_globs.get(parent)
-            if not globs:
-                continue
-            name = sep.join(tokens[-i:])
-            for glob in globs:
-                if self.fnmatch(name, glob, sep=sep):
-                    return True
-        return False
-
+class TarIgnore(GitIgnore):
     def __call__(self, tarinfo: tarfile.TarInfo) -> Optional[tarfile.TarInfo]:
         if self.ignore(tarinfo.path):
             return None
-        abspath = self.root_dir / tarinfo.path
-        if abspath.is_dir():
-            for ignore_file in self.ignore_files:
-                ignore_file = abspath / ignore_file
-                if ignore_file.exists():
-                    self.load_ignore_file(tarinfo.path, ignore_file)
         return tarinfo
